+++ lib/dialog/mutable_value.dart
@@ -1,0 +1,5 @@
+class MutableValue<T> {
+  T value;
+
+  MutableValue(this.value);
+}